requirements.txt
uvicorn==0.24.0.post1
python-socketio==5.11.0
pydantic==2.5.2
fastapi==0.104.1